requirements.txt
absl-py==0.9.0
argh==0.26.2
astor==0.8.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
configparser==4.0.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
docker-pycreds==0.4.0
future==0.18.2
gast==0.3.2
gitdb2==2.0.6
GitPython==3.0.5
glfw==1.8.7
google-pasta==0.1.8
gql==0.1.0
graphql-core==2.2.1
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.4.1
imageio-ffmpeg==0.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
libtmux==0.8.2
lockfile==0.12.2
Markdown==3.1.1
matplotlib
moviepy==0.2.3.5
numpy==1.17.4
nvidia-ml-py3==7.352.0
patchelf
pathtools==0.1.2
Pillow==6.2.1
proglog==0.1.9
promise==2.3
protobuf==3.20.3
psutil==5.6.7
pycparser==2.19
pyglet==1.4.8
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
Rx==1.6.1
scipy==1.4.0
sentry-sdk==0.13.5
shortuuid==0.5.0
six==1.13.0
smmap2==2.0.5
subprocess32==3.5.4
tensorboardX==2.0
termcolor==1.1.0
torch
torchvision
tqdm==4.40.2
urllib3==1.25.7
wandb==0.15.10
watchdog==0.9.0
Werkzeug==0.16.0
wrapt==1.11.2
gym==0.17.1
seaborn==0.10.1
ray==0.8.5
pathfinding==0.0.4
scikit-image
ipdb
gym-minigrid
gdown
attrs
opencv-python
typing
typing_extensions
tensorboard
dcargs
omegaconf==2.3.0
hydra-core==1.3.2
scikit-learn